backend/requirements.txt
Django==4.2.7
pandas
scikit-learn
matplotlib
seaborn
numpy
joblib
gunicorn
whitenoise
python-dotenv
djangorestframework
django-cors-headers
reportlab